Expanded subdirectories, created foundation for autopilot and global control, reorganized code to name each top level directory and .h/.cpp files in PascalCase

+++ lib/global/math/pid/pid.h
@@ -1,19 +1,20 @@
 #pragma once
 
+// WARNING: NOT FINISHED!
 class pid
 {
 public:
-    float compute(float input, float target);
-    pid(float KP, float KI, float KD);
+    double compute(double input, double target);
+    pid(double KP, double KI, double KD);
 
 private:
-    float kP;
-    float kI;
-    float kD;
-    float error;
+    double kP;
+    double kI;
+    double kD;
+    double error;
     unsigned long now;
-    float prevError;
-    float prevTime;
-    float integral;
-    float derivative;
+    double prevError;
+    double prevTime;
+    double integral;
+    double derivative;
 };
